Put RequestError in a separate file to avoid circular dependency
